Fix visual glitch. During the 100ms beep duration, the LED display was frozen on one row of the old number. This glitch was more readily apparent with increasing beep durations.
The fix is to use a PWM - Pulse Width Modulation - to control the speaker, rather than the CPU turning voltage on and off. The API for PWM use in Rust is not well-known and mentioned, but not attempted for being too difficult, in the discovery2 Embedded Rust book. I was lucky to have vibe-coded it correctly.

Now that the speaker and the LED rendering (which works by flickering on the lights at high speeds, one row at a time) are no longer competing for CPU time, they can run concurrently. I reduced the time taken in the critical section of the GPIOTE (reacts to button presses) handler by moving the beeping function out of it (100 ms shaved off), and made the LED rendering preempt the GPIOTE by assigning it a higher priority.